+++ swiftemulator/emulators/gaussian_process_one_dim.py
@@ -177,7 +177,7 @@ class GaussianProcessEmulator1D(BaseEmulator):
         if self.kernel is None:
             number_of_kernel_dimensions = model_specification.number_of_parameters
 
-            self.kernel = 1 ** 2 * george.kernels.ExpSquaredKernel(
+            self.kernel = 1**2 * george.kernels.ExpSquaredKernel(
                 np.ones(number_of_kernel_dimensions), ndim=number_of_kernel_dimensions
             )
 
