### 2022-03-10 #4240 @editorialbot
[REVIEW]: swift-emulator: A Python package for emulation of simulated scaling relations
**Submitting author:** @Moyoxkit (<a href="http://orcid.org/0000-0003-0862-8639">Roi Kugel</a>)
**Repository:** https://github.com/SWIFTSIM/emulator
**Branch with paper.md** (empty if default branch):
**Version:** v1.2.0
