### 2022-03-15 120fc07
Add installation to docs
