### 2022-03-16 836274c
Add velociraptor requirement to setup.py
