### 2022-03-16 ac1a398
Add pyDOE to setup.py
