### 2022-03-17 247e740
Add github workflow without requirements.txt
