### 2022-03-17 324844b
Added emcee to setup.py
