### 2022-03-17 SWIFTSIM/emulator#33 @JDonaldM
[JOSS review] Possible bug in plot_model
Part of [JOSS review](https://github.com/openjournals/joss-reviews/issues/4240).

I have been going through the [Analysis tools examples](https://swiftemulator.readthedocs.io/en/latest/emulator_analysis/index.html) and encountered the following error:
```
Input In [9], in <cell line: 1>()
----> 1 model_parameters.plot_model(model_specification)

File ~/anaconda3/envs/JOSSrev3/lib/python3.8/site-packages/swiftemulator/backend/model_parameters.py:190, in ModelParameters.plot_model(self, model_specification, filename, corner_kwargs)
    187 if corner_kwargs is None:
    188     corner_kwargs = {}
--> 190 corner.corner(
    191     xs=corner_data,
    192     labels=model_specification.parameter_printable_names,
    193     range=model_specification.parameter_limits,
    194     plot_contours=False,
    195     plot_datapoints=True,
    196     plot_density=False,
    197     data_kwargs={"alpha": 1.0},
    198     **corner_kwargs,
    199 )
    201 if filename is None:
    202     plt.show()

TypeError: corner() missing 1 required positional argument: 'data'
```

I can achieve the desired plot is I change line 191 to pass `corner_data` as a positional argument rather than a keyword argument, but I'm not familiar enough with `corner` to know if there is a reason for passing `corner_data` as a keyword argument.

P.S. apologies if this issue appears more than once, there seemed to be an issue with GitHub service today!
