### 2022-03-17 SWIFTSIM/emulator#34 @JDonaldM
[JOSS review] Minor comments about examples in the docs.
[Review thread](https://github.com/openjournals/joss-reviews/issues/4240).

I haven't looked at at all of the examples in the docs yet but the ones I have looked at have been really nice and give a good overview of what you can do with the code!

I did want to note that in a few cases the code blocks cannot be run in the order they are presented without a couple of tweaks. For example:

- The [analysis tools](https://swiftemulator.readthedocs.io/en/latest/emulator_analysis/index.html) examples rely on the user to have read the [getting started](https://swiftemulator.readthedocs.io/en/latest/getting_started/index.html) example first and realise that the 'The example data will be the Schecter function example' refers to this. In the same example the cross check plot uses `matplotlib` but it isn't imported in any of the previous code blocks, also the parameter sweep plot calls `schecter_emulator` which isn't defined in this example.
- In the [experimental design](https://swiftemulator.readthedocs.io/en/latest/swift_io/design.html) example the final code block throws an error because of string formatting. It might also be good to link to `eagle_25.yml`, it wasn't immediately clear to me where to find this file.
- In the [loading SWIFT data](https://swiftemulator.readthedocs.io/en/latest/swift_io/data_loading.html) example `files` and `value_files` in the first code block are undefined, along with `filenames` in the third code block.
