### 2022-03-18 SWIFTSIM/emulator#33 @Moyoxkit
Hey @JDonaldM,

Thanks for the comment, this was indeed a problem. the positional argument "xs" stems from an older version of corner and has since then been replaced by "data". In order to make the plotting work with both versions, it is now passed as a positional argument.

This should now be fixed in the latest master.
