### 2022-03-18 a52d2b1
Update swift example
