### 2022-03-18 SWIFTSIM/emulator#34 @Moyoxkit
Hello @JDonaldM

Thanks for the comments! It is indeed much nicer to have each page be self-consistent. I've implemented updates to these sections here: https://github.com/SWIFTSIM/emulator/pull/35 We'll merge the changes to the master once https://github.com/SWIFTSIM/emulator/issues/22 is also fixed.
