### 2022-03-21 ca20fc2
Fix the cornerplots in the swift example
