### 2022-03-21 SWIFTSIM/emulator#34 @Moyoxkit
Have now been resolved with #35 and #36
