### 2022-03-21 7b9dbc0
Updatae emulator options to be self contained
