### 2022-03-23 SWIFTSIM/emulator#37 @kstoreyf
[JOSS review] small fixes in Interaction with SWIFT examples
Part of [JOSS review](https://github.com/openjournals/joss-reviews/issues/4240).

Hello, I was running the example code for the Interaction with SWIFT examples and hit a few small bugs / things that could be more self-consistent.

1. On the [Experimental Design]((https://swiftemulator.readthedocs.io/en/latest/swift_io/design.html) page, in the last code block,

I believe
`key: output_path / f"{key}.yml"`
should be something like
`key: f"{output_path}/{key}.yml"`
?

Just saw that this was referenced in #34 but I don't think it has been fixed.

2. For the example on [Loading SWIFT Data](https://swiftemulator.readthedocs.io/en/latest/swift_io/data_loading.html#), it uses the `load_pipeline_outputs` function without importing it on the page. It would be good to include the line
`from swiftemulator.io.swift import load_pipeline_outputs`

3. Similarly, in the last code block on [Loading SWIFT Data](https://swiftemulator.readthedocs.io/en/latest/swift_io/data_loading.html#), replace `emulator = GaussianProcessEmulator()` with
```
from swiftemulator.emulators import gaussian_process
emulator = gaussian_process.GaussianProcessEmulator()
```

Thanks!
