### 2022-03-23 SWIFTSIM/emulator#38 @kstoreyf
[JOSS review] Minor fixes in docs
Part of [JOSS review](https://github.com/openjournals/joss-reviews/issues/4240).

The examples are generally very clean and detailed, nicely done! I have a few small things I noticed while I was going through them:

1. For the velociraptor comparison, it is not clear how to get the data that is used in the example code on that page. The page links to the [velociraptor-comparison-data](https://github.com/SWIFTSIM/velociraptor-comparison-data/blob/master/convert.py) repo, but it would be nice to give the steps for downloading the `Vernon.hdf5` file needed into the right place (requires cloning the repo and running a conversion script, which I figured out after a bit but wasn't obvious). Alternatively, you could provide a direct link to download that file.
2. ReadTheDocs isn't properly pulling in the data for the [visualization module](https://swiftemulator.readthedocs.io/en/latest/modules/swiftemulator.comparison.visualisation.html).
3. It would be nice to add axis labels (and legends where applicable) to the plots in the examples, so they can be understood without having to read the plotting code, e.g. on the Getting Started page, the Sweep analysis tool, and the penalty plots.
