### 2022-03-24 SWIFTSIM/emulator#39 @kstoreyf
[JOSS review] Comments on paper
Part of [JOSS review](https://github.com/openjournals/joss-reviews/issues/4240).

The paper contains all of the required information, and is clear and well-written.  The authors make a strong case for the utility of this emulation framework, situated between low-level packages and specific applications - indeed, I couldn't find another framework quite like it!

I have a few comments / minor fixes:

1. For the emulation formula, it would be good to also define the dependent variable *y*. This is done on the [Getting Started](https://swiftemulator.readthedocs.io/en/latest/getting_started/index.html) page - in fact, it might be nice to have consistent notation between the formula here and the one in the paper (one way or the other).
2. I don't really understand what the line "This distinction between the values obtained from the simulation and the parameters of the simulation model is absent when setting up a typical emulator” (lines 51-52) is trying to say. Could you elaborate, or use more precise language?
3. Lines 28-29: Put each citation directly after the simulation project it is for.
4. Small typos:
- Line 23: state of the of the art -> state-of-the-art
- Line 62: Many packages cexist -> exist
- Line 85: ranging accros -> across
5. I am wondering about consistency in the name of the emulator. In the paper it is titled swift-emulator, the repo is emulator, the pip install is swiftemulator, and the docs call it both SWIFT-Emulator and SWIFT emulator. I understand that some of these are set by external constraints, e.g. pip not allowing hyphens, and it is usually still clear, but it might be good to have a bit more consistency where possible.
