+++ docs/source/emulator_analysis/index.rst
@@ -140,8 +140,10 @@ This is implemented into the SWIFT-Emulator with
     for mock_name in Mock_values.keys():
         plt.plot(Mock_values[mock_name]["independent"],
                 Mock_values[mock_name]["dependent"],
-                label = str(Mock_parameters[mock_name]["alpha"])[:4])
-        
+                label = "Alpha = " +str(Mock_parameters[mock_name]["alpha"])[:4])
+
+    plt.xlabel("Stellar mass")
+    plt.ylabel("dn/dlogM")    
     plt.legend()
     plt.savefig("parameter_sweep.png",dpi=200)
 
