+++ docs/source/emulator_analysis/index.rst
@@ -135,7 +135,7 @@ This is implemented into the SWIFT-Emulator with
 
     Mock_values, Mock_parameters = mock_sweep(schecter_emulator
                                           ,model_specification
-                                          ,10,"alpha",center)
+                                          ,6,"alpha",center)
 
     for mock_name in Mock_values.keys():
         plt.plot(Mock_values[mock_name]["independent"],
