### 2022-03-24 SWIFTSIM/emulator#37 @Moyoxkit
Thanks for the comments, will be fixed with #40
