### 2022-03-24 67aa243
Make thetas vectors again
