### 2022-03-28 SWIFTSIM/emulator#39 @Moyoxkit
Thanks for the comments! Changes should be merged now with #40
