### 2022-03-28 33df21a
Minor changes in vis to help docs work
Signed-off-by: Josh Borrow <josh@joshborrow.com>
