+++ swiftemulator/emulators/gaussian_process_bins.py
@@ -213,7 +213,7 @@ class GaussianProcessEmulatorBins(BaseEmulator):
                     self.model_specification.number_of_parameters
                 )
 
-                self.kernel = 1**2 * george.kernels.ExpSquaredKernel(
+                self.kernel = 1 ** 2 * george.kernels.ExpSquaredKernel(
                     np.ones(number_of_kernel_dimensions),
                     ndim=number_of_kernel_dimensions,
                 )
