+++ swiftemulator/emulators/gaussian_process_mcmc.py
@@ -210,7 +210,7 @@ class GaussianProcessEmulatorMCMC(BaseEmulator):
                 self.model_specification.number_of_parameters + 1
             )
 
-            self.kernel = 1**2 * george.kernels.ExpSquaredKernel(
+            self.kernel = 1 ** 2 * george.kernels.ExpSquaredKernel(
                 np.ones(number_of_kernel_dimensions), ndim=number_of_kernel_dimensions
             )
 
