### 2022-03-28 #4240 @JBorrow
I've pushed a new version of the package to PyPI with fixes to all of the attached issues. Thanks both for all your constructive comments!
