### 2022-03-30 SWIFTSIM/emulator#37 @kstoreyf
Thanks! For 3, I think you now need to call `gaussian_process.GaussianProcessEmulator()`, not just `GaussianProcessEmulator()`, after the import statement.
