### 2022-03-31 SWIFTSIM/emulator#37 @Moyoxkit
Yes indeed, has been fixed now
