### 2022-04-08 958f4c2
Update citation formatting
