+++ paper/paper.bib
@@ -15,20 +15,16 @@
   adsnote       = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@article{Gardner2018,
-  author        = {{Gardner}, Jacob R. and {Pleiss}, Geoff and {Bindel}, David and {Weinberger}, Kilian Q. and {Wilson}, Andrew Gordon},
-  title         = {{GPyTorch: Blackbox Matrix-Matrix Gaussian Process Inference with GPU Acceleration}},
-  journal       = {arXiv e-prints},
-  keywords      = {Computer Science - Machine Learning, Statistics - Machine Learning},
-  year          = 2018,
-  month         = sep,
-  eid           = {arXiv:1809.11165},
-  pages         = {arXiv:1809.11165},
-  archiveprefix = {arXiv},
-  eprint        = {1809.11165},
-  primaryclass  = {cs.LG},
-  adsurl        = {https://ui.adsabs.harvard.edu/abs/2018arXiv180911165G},
-  adsnote       = {Provided by the SAO/NASA Astrophysics Data System}
+@inproceedings{Gardner2018,
+ author = {Gardner, Jacob and Pleiss, Geoff and Weinberger, Kilian Q and Bindel, David and Wilson, Andrew G},
+ booktitle = {Advances in Neural Information Processing Systems},
+ editor = {S. Bengio and H. Wallach and H. Larochelle and K. Grauman and N. Cesa-Bianchi and R. Garnett},
+ pages = {},
+ publisher = {Curran Associates, Inc.},
+ title = {{GPyTorch}: Blackbox Matrix-Matrix {G}aussian Process Inference with {GPU} Acceleration},
+ url = {https://proceedings.neurips.cc/paper/2018/file/27e8e17134dd7083b050476733207ea1-Paper.pdf},
+ volume = {31},
+ year = {2018}
 }
 
 @misc{Elahi2019,
@@ -78,7 +74,7 @@
 @article{Vogelsberger2014,
   author        = {{Vogelsberger}, Mark and {Genel}, Shy and {Springel}, Volker and {Torrey}, Paul and {Sijacki}, Debora and {Xu}, Dandan and {Snyder}, Greg and {Nelson}, Dylan and {Hernquist}, Lars},
   title         = {{Introducing the Illustris Project: simulating the coevolution of dark and visible matter in the Universe}},
-  journal       = {\mnras},
+  journal       = {Monthly Notices of the Royal Astronomical Society},
   keywords      = {methods: numerical, cosmology: theory, Astrophysics - Cosmology and Nongalactic Astrophysics},
   year          = 2014,
   month         = oct,
@@ -96,7 +92,7 @@
 @article{Crain2015,
   author        = {{Crain}, Robert A. and {Schaye}, Joop and {Bower}, Richard G. and {Furlong}, Michelle and {Schaller}, Matthieu and {Theuns}, Tom and {Dalla Vecchia}, Claudio and {Frenk}, Carlos S. and {McCarthy}, Ian G. and {Helly}, John C. and {Jenkins}, Adrian and {Rosas-Guevara}, Yetli M. and {White}, Simon D.~M. and {Trayford}, James W.},
   title         = {{The EAGLE simulations of galaxy formation: calibration of subgrid physics and model variations}},
-  journal       = {\mnras},
+  journal       = {Monthly Notices of the Royal Astronomical Society},
   keywords      = {galaxies: evolution, galaxies: formation, galaxies: haloes, cosmology: theory, Astrophysics - Astrophysics of Galaxies},
   year          = 2015,
   month         = jun,
@@ -114,7 +110,7 @@
 @article{Pillepich2018,
[…]
