+++ paper/paper.bib
@@ -27,7 +27,9 @@
  year = {2018}
 }
 
-@misc{Elahi2019,
+@article{Elahi2019,
+  journal = {The Astrophysics Source Code Library},
+  url = {http://ascl.net/1911.020},
   author        = {{Elahi}, Pascal Jahan and {Poulton}, Rhys and {Canas}, Rodrigo},
   title         = {{VELOCIraptor-STF: Six-dimensional Friends-of-Friends phase space halo finder}},
   keywords      = {Software},
@@ -41,7 +43,9 @@
   adsnote       = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@misc{Schaller2018,
+@article{Schaller2018,
+  journal = {The Astrophysics Source Code Library},
+  url = {http://ascl.net/1805.020},
   author        = {{Schaller}, Matthieu and {Gonnet}, Pedro and {Draper}, Peter W. and {Chalk}, Aidan B.~G. and {Bower}, Richard G. and {Willis}, James and {Hausammann}, Lo{\"\i}c},
   title         = {{SWIFT: SPH With Inter-dependent Fine-grained Tasking}},
   keywords      = {Software},
