+++ examples/basic_example.py
@@ -24,7 +24,7 @@ def my_true_model(x, offset):
     """
     A basic non-linear model.
     """
-    return offset + offset ** 2 * x ** 0.5
+    return offset + offset**2 * x**0.5
 
 
 model_specification = se.ModelSpecification(
