### 2022-04-11 #4240 @JBorrow
Yes, we have pushed 1.2.0 to include the final changes to the code required for JOSS acceptance.
