### 2022-04-12 #4240 @arfon
@JBorrow – I spotted a couple of minor issues in your paper which I've fixed here: https://github.com/SWIFTSIM/emulator/pull/43
